+++ packages/react/src/components/Button/index.stories.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Story } from '@storybook/react';
-import { Button, IconButton, ButtonProps } from '.';
+import { Button, IconButton } from '.';
 import { IconButtonProps } from './IconButton';
 import { IconOptions } from '../Icon';
+import { ButtonProps } from './types';
 
 export default {
 	title: 'Button',
